Update references to team ODS code
The ODS code no longer belongs on each team as some times will share the
same ODS code in the future. Instead, the ODS code can be found on the
related organisation model.

For now, there's no longer a unique identifier on each team. We will
need to add one to support logging in where we need to know which team
a user belongs to through their CIS2 workgroup.

At the moment we assume that each organisation only ever has one team,
but this won't always be the case and needs to be supported in a follow
up pull request.

Jira-Issue: MAV-1280