Fix issue with incorrect patients appearing in programmes
This fixes a bug where patients would appear in the wrong programmes for
their year groups. The reason this was happening is because the
resulting SQL query generated was incorrect:

- `PatientSession#appear_in_programmes` scope joined on the patients.
- `Patient#appear_in_programmes` scope used the scope on
  `PatientSession` but it also did a `where("patient_id =
  patients.id")`.
- Due to these to facts, in the resulting SQL, the `patient_id` was
  referring to `patient_sessions.id` and so was `patients.id` (due
  to the join), meaning it would also be `true`.

Therefore, this scope would only work if there was only 1 patient. If
not, any patient sessions for any other patients would be caught in the
`EXISTS`.

I've added a unit test for the scope to cover this scenario and added
more to the feature test to ensure we don't regress.